Moved DoneButton component to components directory
Related to https://github.com/Automattic/wp-calypso/pull/59178#issuecomment-994091221
